projects/web_page_summarizer/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
openai==1.60.2
rich==13.9.4